Use wrapper for RDV w/ bugfix instead of upstream RDV directly
