+++ changelogs/unreleased/th__use_rdv_wrapper.yaml
@@ -1,0 +1,2 @@
+fixed:
+  - array and pod scalarization passes used the original RDV pass instead of the wrapper with bug fix
