+++ lib/Dialect/Array/Transforms/ArrayToScalarPass.cpp
@@ -969,7 +969,7 @@ class PassImpl : public llzk::array::impl::ArrayToScalarPassBase<PassImpl> {
         }
     ));
     // Cleanup SSA values made dead by removing allocations and writes.
-    nestedPM.addPass(createRemoveDeadValuesPass());
+    nestedPM.addPass(createRemoveDeadValuesWorkaroundPass());
     if (failed(runPipeline(nestedPM, module))) {
       signalPassFailure();
       return;
