+++ tools/llzk-opt/llzk-opt.cpp
@@ -31,6 +31,7 @@
 #include "llzk/Dialect/POD/Transforms/TransformationPasses.h"
 #include "llzk/Dialect/Polymorphic/Transforms/TransformationPasses.h"
 #include "llzk/Transforms/LLZKTransformationPasses.h"
+#include "llzk/Transforms/SpecializedMemoryPasses.h"
 #include "llzk/Validators/LLZKValidationPasses.h"
 
 #include <mlir/Dialect/Func/Extensions/InlinerExtension.h>
@@ -68,53 +69,6 @@ static llvm::cl::opt<bool>
 /// encounters an `scf.if` op with an empty else region.
 namespace mlir_hotfix {
 
-class RemoveDeadValuesWorkaroundPass
-    : public mlir::PassWrapper<RemoveDeadValuesWorkaroundPass, mlir::OperationPass<>> {
-public:
-  MLIR_DEFINE_EXPLICIT_INTERNAL_INLINE_TYPE_ID(RemoveDeadValuesWorkaroundPass)
-
-  llvm::StringRef getArgument() const override { return "remove-dead-values"; }
-  llvm::StringRef getDescription() const override { return "Remove dead values"; }
-
-  void runOnOperation() final {
-    // pre-pass: add trivial block to empty "else" regions of scf.if ops
-    getOperation()->walk([](mlir::scf::IfOp ifOp) {
-      if (ifOp.getElseRegion().empty()) {
-        mlir::Block &elseBlock = ifOp.getElseRegion().emplaceBlock();
-        mlir::OpBuilder builder(ifOp.getContext());
-        builder.setInsertionPointToEnd(&elseBlock);
-        builder.create<mlir::scf::YieldOp>(ifOp.getLoc());
-      }
-    });
-
-    mlir::OpPassManager pm(getOperation()->getName().getStringRef());
-    pm.addPass(mlir::createRemoveDeadValuesPass());
-    if (mlir::failed(runPipeline(pm, getOperation()))) {
-      signalPassFailure();
-    }
-
-    // post-pass: cleanup trival "else" blocks that remain after the pass
-    getOperation()->walk([](mlir::scf::IfOp ifOp) {
-      if (ifOp.getResults().empty()) {
-        mlir::Region &elseRegion = ifOp.getElseRegion();
-        if (!llvm::hasSingleElement(elseRegion)) {
-          return;
-        }
-        mlir::Block &elseBlock = elseRegion.front();
-        if (!llvm::hasSingleElement(elseBlock)) {
-          return;
-        }
-        if (!llvm::isa<mlir::scf::YieldOp>(elseBlock.front())) {
-          return;
-        }
-        elseRegion.dropAllReferences();
-        elseBlock.clear();
-        elseRegion.getBlocks().clear();
-      }
-    });
-  }
-};
-
 inline static void registerTransformsPasses() {
   mlir::registerCSE();
   mlir::registerCanonicalizer();
@@ -129,7 +83,7 @@ inline static void registerTransformsPasses() {
   mlir::registerPrintIRPass();
   mlir::registerPrintOpStats();
   mlir::registerPass([]() -> std::unique_ptr<mlir::Pass> {
-    return std::make_unique<RemoveDeadValuesWorkaroundPass>();
+    return llzk::createRemoveDeadValuesWorkaroundPass();
   });
   mlir::registerSCCP();
   mlir::registerSROA();
